+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Added:
 - `timeoutToGetAnUiElement` to the config
 - Documentation!
 - `Assertion.Checked` and `Assertion.NotChecked` to check if a checkbox/switch is checked or not
+- `Assertion.Enabled` and `Assertion.NotEnabled` to check if a view is enabled or not
 
 Changed:
 - Filter out empty children with no `Attribute.TEXT` or `Attribute.RESOURCE_ID` in ViewHierarchy